requirements.txt
transformers
flask
python-dotenv
marshmallow
gunicorn
Flask-Smorest
onnx==1.14.1
onnxruntime==1.15.1
optimum[onnxruntime]==1.14.1
watchdog
